config(nvim): do not lazy load mason
* Breaks LSP initialization

Mason adds LSP executable bin path to NEOVIM path on require("mason").setup() however,
lazy loading defers this behavior until `Mason` command is ran. Because of this LSP
executables paths won't be available for the vim.lsp to spawn process for buffers.

Hence the error: *LSP NAME* is not an executable